+++ src/settings/settings-tab.ts
@@ -42,5 +42,15 @@ export class AutocompleteSettingsTab extends PluginSettingTab {
           this.plugin.refresh()
         })
       )
+    new Setting(containerEl)
+      .setName('Flow Provider')
+      .setDesc('Learns as you type. For now limited to current session.')
+      .addToggle((cb) =>
+        cb.setValue(this.plugin.settings.latexProvider).onChange((value) => {
+          this.plugin.settings.flowProvider = value
+          this.plugin.saveData(this.plugin.settings)
+          this.plugin.refresh()
+        })
+      )
   }
 }