+++ src/autocomplete.ts
@@ -1,14 +1,17 @@
 import {
   Direction,
   defaultDirection,
+  completionWordIn,
+  managePlaceholders,
+} from './autocomplete/core'
+import {
   generateView,
-  getRange,
   appendWidget,
   updateCachedView,
   scrollTo,
-  completionWordIn,
 } from './autocomplete/view'
-import LatexProvider from './providers/latex'
+import { FlowProvider } from './providers/flow'
+import LaTexProvider from './providers/latex'
 import { Completion, Provider } from './providers/provider'
 import { AutocompleteSettings } from './settings/settings'
 
@@ -63,7 +66,7 @@ export class Autocomplete {
 
   public removeViewFrom(editor: CodeMirror.Editor) {
     this.selected = defaultDirection()
-    this.removeKeyBindings(editor)
+    editor.removeKeyMap(this.keyMaps)
 
     if (!this.view) return
     this.addClickListener(this.view, editor, false)
@@ -79,6 +82,18 @@ export class Autocomplete {
     }
   }
 
+  public updateProvidersFrom(event: KeyboardEvent, editor: CodeMirror.Editor) {
+    if (!event.ctrlKey && !event.altKey && event.key === ' ') {
+      const cursor = editor.getCursor()
+      const line = editor.getLine(cursor.line)
+      this.providers.forEach((provider) => {
+        // For now only FlowProvider
+        if (provider instanceof FlowProvider)
+          provider.addCompletionWord(line, cursor.ch - 1)
+      })
+    }
+  }
+
   private showViewIn(editor: CodeMirror.Editor, completionWord: string = '') {
     if (this.view) this.removeViewFrom(editor)
 
@@ -87,7 +102,7 @@ export class Autocomplete {
       []
     )
 
-    this.addKeyBindings(editor)
+    editor.addKeyMap(this.keyMaps)
 
     this.view = generateView(this.suggestions, this.selected.index)
     this.addClickListener(this.view, editor)
@@ -113,14 +128,6 @@ export class Autocomplete {
         }
         break
     }
-  }
-
-  private addKeyBindings(editor: CodeMirror.Editor) {
-    editor.addKeyMap(this.keyMaps)
-  }
-
-  private removeKeyBindings(editor: CodeMirror.Editor) {
-    editor.removeKeyMap(this.keyMaps)
   }
 
   private keyMaps = {
@@ -177,12 +184,15 @@ export class Autocomplete {
   private selectSuggestion(editor: CodeMirror.Editor) {
     const cursor = editor.getCursor()
     const selectedValue = this.suggestions[this.selected.index].value
-    const [replaceFrom, replaceTo] = getRange(this.cursorAtTrigger, cursor.ch)
+
+    const { normalizedValue, newCursorPosition } = managePlaceholders(
+      selectedValue,
+      this.cursorAtTrigger!.ch
+    )
 
     editor.operation(() => {
-      editor.replaceRange(selectedValue, replaceFrom, replaceTo)
+      editor.replaceRange(normalizedValue, this.cursorAtTrigger, cursor)
 
-      const newCursorPosition = replaceFrom.ch + selectedValue.length
       const updatedCursor = {
         line: cursor.line,
         ch: newCursorPosition,
@@ -196,7 +206,8 @@ export class Autocomplete {
 
   private loadProviders() {
     const providers = []
-    if (this.settings.latexProvider) providers.push(new LatexProvider())
+    if (this.settings.latexProvider) providers.push(new LaTexProvider())
+    if (this.settings.flowProvider) providers.push(new FlowProvider())
 
     this.providers = providers
   }