+++ src/providers/provider.ts
@@ -1,6 +1,8 @@
 export abstract class Provider {
   abstract readonly category: string
-  abstract readonly completions: Array<string>
+  abstract completions: Array<string>
+
+  static placeholder: string = '#{}'
 
   matchWith(input: string): Completion[] {
     // TODO: Improve filtering with weights