+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.6.0
+- Move cursor into placeholder, if any
+- Updated features plan in readme
+- Added Flow provider
+- Updated gif in readme
+
 # 0.5.10
 - Forgot to update version files 🤦